requirements.txt
asgiref==3.7.2
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.0
beautifulsoup4==4.12.2
boto3==1.28.30
botocore==1.31.30
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.6.2
colorama==0.4.5
cryptography==38.0.1
distro==1.7.0
Django==4.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==24.3
django-storages==1.13.2
djangorestframework==3.14.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
emoji==2.8.0
emot==3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==3.2.0
Markdown==3.4.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.23.4
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
pandas==1.5.1
paramiko==2.11.0
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pywin32==304
PyYAML==5.4.1
requests==2.28.1
s3transfer==0.6.2
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
texttable==1.6.4
toml==0.10.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.3.0
tzdata==2022.2
unicode==2.9
uritemplate==4.1.1
urllib3==1.26.12
vboxapi==1.0
websocket-client==0.59.0
wsproto==1.2.0
